Hack-fixed setup.py after some of its black magic ceased to work. It should be rewritten since no one understands the code well enough to be able to maintain it.

+++ setup.py
@@ -71,6 +71,10 @@ class _build_i18n(distutils.cmd.Command):
 		"""
 		Update the language files, generate mo files and add them
 		to the to be installed files
+
+		NOTE: This code is partly broken and hack-fixed to the state where it appears to work.
+		      It should be removed, since nobody understands the code well enough to be able to maintain it.
+
 		"""
 		if not os.path.isdir(self.po_dir):
 			return
@@ -149,10 +153,15 @@ class _build_i18n(distutils.cmd.Command):
 
 		# Since specifying a .mofile dir is not supported, we manually move build/mo/
 		# to a place more appropriate in our opinion, currently content/lang/.
-		if os.path.exists(os.path.join("content", "lang")):
-			rmtree(os.path.join("content", "lang"))
-		copytree(os.path.join("build", "mo"), \
-			os.path.join("content", "lang"))
+
+		if os.path.exists(os.path.join("build", "mo")):
+			# it appears build/mo should always magically appear, but does not on some gentoo machines.
+			# there, everything is placed in content/lang, so it's fine
+			# on other machines, we have to move stuff around like that:
+			if os.path.exists(os.path.join("content", "lang")):
+				rmtree(os.path.join("content", "lang"))
+			copytree(os.path.join("build", "mo"), \
+				os.path.join("content", "lang"))
 
 build.sub_commands.append(('build_i18n', None))
